refactor(meeting): extract AppClassifier to its own peer (#431 item 2)
Pure code-move refactor. The 2,193-LOC `meeting/manager.rs` had
~180 LOC of `AppClassifier` (bundle-id → MeetingAppKind lookup +
defaults table + override layering) sitting alongside the lifecycle
state machine — a different concern that's purely declarative and
entirely independent of the manager's mutable state.

Lifts `AppClassifier` (struct + impl + the curated defaults table) into
`meeting/classifier.rs`, mirroring the pattern existing peer modules
(`autostart.rs`, `autostart_poller.rs`, `app_overrides.rs`,
`pump.rs`) already use. Re-exported from `meeting::mod.rs` so the
public API (`crate::meeting::AppClassifier`) is unchanged.
`autostart_poller.rs` updates its internal import from
`super::manager::AppClassifier` to `super::classifier::AppClassifier`.

`manager.rs` drops to 2,011 LOC.

The audit's other recommendation — lifting `start_manual` /
`stop_manual` / `append_if_active` into a `lifecycle.rs` peer —
is deferred. Those methods touch ~10 private fields of
`SessionManager` and the private `ActiveSession` / `SessionState`
types, so the split needs `pub(super)` visibility on every field
in the lifecycle hot path. That's a separate, more careful PR
with reviewer eyes on the surface-area change, not a sweep.

372 lib tests pass, clippy `-D warnings` clean, npm run check
clean, 87 Playwright specs pass.
